### 2023-10-15 #5950 @editorialbot
[PRE REVIEW]: '`cstag` and `cstag-cli`: tools for manipulating and visualizing CS tags'
**Submitting author:** @akikuno (<a href="http://orcid.org/0000-0002-4674-6882">Akihiro Kuno</a>)
**Repository:** https://github.com/akikuno/cstag
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.4
