### 2023-10-15 #5950 @Kevin-Mattheus-Moerman
@akikuno thanks for this submission. I am the AEiC for this track, and here to help with initial steps. Note I have just triggered a scope review by our editorial board. This scope review is to help me determine if this work conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), e.g. in terms of size/maturity/functionality.

This scope review should take about 2 weeks to complete.
