### 2023-10-15 #5950 @Kevin-Mattheus-Moerman
Note to editors. This looks like this submission also includes: https://github.com/akikuno/cstag-cli, for which `cloc` gives us:
```
Language                     files          blank        comment           code
Markdown                         4             69              0            166
Python                           4             40             18             90
YAML                             3             14              0             68
Bourne Shell                     1             18             18             33
TOML                             1              4              0             27
SUM:                            13            145             36            384
```
