### 2023-11-02 #5950 @Kevin-Mattheus-Moerman
@akikuno I have just removed the `query-scope` label as I am pleased to inform you this work has passed the initial scope check by our editorial board. I will now assign a handling editor. Note though that the editor I'll assign (@jmschrei) is handling many other submissions and expects to be able to get to this submission within about 2 weeks.

If you'd like to suggest reviewers at this point that would be helpful for us. You can do so by listing their GitHub handles, but please do leave out the `@` symbol so they are not tagged.

@akikuno on your repository, I see you have a code-of-conduct, and the README talks about "Feedback and Support" (and posting issues), however, can  you also add instructions for users on how to contribute? A detailed way of doing so would be to add a `CONTRIBUTING.md` file (e.g. see also examples here: https://contributing.md/example/), a more basic approach is to simply spell out some simple instructions in the README.
