+++ pyproject.toml
@@ -21,5 +21,5 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8"
 weasyprint = ">=52.5"
