+++ .github/workflows/pytest.yml
@@ -1,14 +1,18 @@
 name: Pytest
 
-on: [push]
+on:
+  push:
+  schedule:
+    - cron: '0 2 * * 1' # The task runs at 2 a.m. every Monday.
 
 jobs:
   build:
     runs-on: ${{ matrix.os }}
     strategy:
+      max-parallel: 4
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: [ '3.8', '3.9', '3.10', '3.x' ]
+        python-version: [ '3.8', '3.9', '3.10', '3.11' ]
     name: Python ${{ matrix.python-version }} ${{ matrix.os }}
 
     steps:
