### 2023-11-03 c0f4dff
add CONTRIBUTING.md
