+++ README.md
@@ -212,6 +212,9 @@ You can obtain the same images of `cstag.to_html` as a PDF file.
 For questions, bug reports, or other forms of feedback, we'd love to hear from you!  
 Please use [GitHub Issues](https://github.com/akikuno/cstag/issues) for all reporting purposes.  
 
+Please refer to [CONTRIBUTING](https://github.com/akikuno/cstag/blob/main/CONTRIBUTING.md) for how to contribute and how to verify your contributions.  
+
+
 ## 🤝 Code of Conduct
 
 Please note that this project is released with a [Contributor Code of Conduct](https://github.com/akikuno/cstag/blob/main/CODE_OF_CONDUCT.md).  
