### 2023-11-03 #5950 @akikuno
@Kevin-Mattheus-Moerman @jmschrei

I appreciate your feedback. I am very pleased to be able to undergo the review process with JOSS.

As recommended, I have included a CONTRIBUTING.md file in the repository, which can be found here:
https://github.com/akikuno/cstag/blob/main/CONTRIBUTING.md

For potential reviewers, I propose the following individuals:
- [Marcus D Sherman](https://github.com/betteridiot), author of [Bamnostic](https://joss.theoj.org/papers/10.21105/joss.00826),
- [Jesse D. Bloom](https://github.com/jbloom), author of [Alignparse](https://joss.theoj.org/papers/10.21105/joss.01915),
- [Kazuki Nakamae](https://github.com/KazukiNakamae), a known expert in the field of genome editing analysis.

Kind regards,
