### 2023-11-17 #6066 @editorialbot
[REVIEW]: '`cstag` and `cstag-cli`: tools for manipulating and visualizing CS tags'
**Submitting author:** @akikuno (<a href="http://orcid.org/0000-0002-4674-6882">Akihiro Kuno</a>)
**Repository:** https://github.com/akikuno/cstag
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.0.5, 1.0.0
