### 2023-11-19 akikuno/cstag#24 @jbloom
is it possible to have the `consensus` function allow some requirements re level of agreement in a consensus?
Not sure if this is feasible to implement, but I wanted to make this request which would make `cstag` more useful to our group.

Right now the `consensus` function doesn't contain any mechanism for checking the "quality" of the consensus. In our work, we often want to try to call a consensus, but also be able to discard sets of reads if the consensus isn't in good agreement. This is common in sequencing of amplicons for many purposes. Eg, we would ideally like:

1. Return the consensus if the individual cs tags agree at at least some level that could be different from majority voting. (eg, maybe we say 75% have to agree). If they don't, provide some indication consensus should be dropped.
2. Similarly be able to report the maximal edit distance of any one tag from the others.
