### 2023-11-19 #6066 @jbloom
I have reviewed the `cstag` submission by @akikuno. I support publication. This is a straightforward and well-documented package that solves a useful question, namely how to add / manipulate `cs` tags. I could definitely see using this package in my own work at some point, and it is well enough documented that I can figure out how I would do that.

I have completed the reviewer checklist above.

I would add the following notes that could be addressed in a few minor revisions:

1. Could be helpful if it was a little bit more clearly indicated that `cstag-cli` is installed from a separate repo.
2. In author could consider citing [alignparse](https://github.com/jbloomlab/alignparse), which provides some similar-ish functionality in its `cs_tag` module.
3. I have made the [following feature request](https://github.com/akikuno/cstag/issues/24). Note that this request is optional and not required for me to recommend acceptance of the current package, but if some similar feature was added it would make `cstag` package more useful for our work.
