### 2023-11-20 akikuno/cstag#25 @betteridiot
[RE: JoSS Review] CS tag deconfliction
For the JoSS review at openjournals/joss-reviews/issues/6066.

The `CS` (uppercase) is a reserved for color read sequence information as per the [SAM format](https://samtools.github.io/hts-specs/SAMtags.pdf), while the `cs` (lowercase)
tag represents the difference string. This has already been a [topic of discussion](https://github.com/samtools/samtools/issues/1030#issuecomment-477137623) while feature requests to incorporate the `cs` tag more
formally into the SAM format.

It might be beneficial to change the "CS" to "cs" in all documentation so as to not mislead users.
