### 2023-11-20 #6066 @betteridiot
Here are my notes for the review. Very minor fixes throughout with a couple of larger issues that will be addressed via GitHub issues on author's issue tracker.

Please see complete notes of the review at [betteridiot/joss_reviews/blob/master/cstag/joss_review.md](https://github.com/betteridiot/joss_reviews/blob/master/cstag/joss_review.md)

# Documentation
1. Paper version references `cstag` v.1.0.4, but the current version available is v1.0.5
2. Documentation does not express statement of need. It only describes what the program does.

# Statement of the field
1. Acknowledges `2passtools` but does not address its internal [cs tag functionality](https://github.com/bartongroup/2passtools/blob/d4378d0f4a780b5b89399161ed8eac6ca5092128/lib2pass/bamparse.py#L13-L56)
2. I am curious about the author's "rebranding" of [`calcs`](https://pypi.org/project/calcs/), which appears to be a precursor to /blob/b0d72945421b361cfed682e2c4d7dfb38ec1842c/alignparse/cs_tag.py) in their state of the field.
3. As addressed by @jbloom, the author should recognize [`alignparse`](https://github.com/jbloomlab/alignparse/blob/b0d72945421b361cfed682e2c4d7dfb38ec1842c/alignparse/cs_tag.py) in their state of the field.

# References
* Missing DOIs for `minimap2` and Bioconda
