### 2023-11-20 akikuno/cstag#26 @betteridiot
[RE: JoSS Review] failing doctests
For the JoSS review at openjournals/joss-reviews/issues/6066

Please see https://github.com/betteridiot/joss_reviews/blob/master/cstag/joss_review.md for complete notes.

Doctests for `cstag` do not pass do to minor syntax errors. See [Testing doctests](https://github.com/betteridiot/joss_reviews/blob/master/cstag/joss_review.md##testing-doctests). As doctests should also be testable (given their nature), this should be addressed.
