### 2023-11-23 4b29ecf
add a reference of alignparse and add DOIs
