+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: '`cstag` and `cstag-cli`: tools for manipulating and visualizing CS tags'
+title: '`cstag` and `cstag-cli`: tools for manipulating and visualizing cs tags'
 tags:
   - python
   - genomics
@@ -21,20 +21,21 @@ bibliography: paper.bib
 
 # Summary
 
-With the widespread commoditization of DNA sequencers, the scientific community is generating an unprecedented volume of sequence alignments on a daily basis. Conventionally, the task of identifying specific mutations or reconstructing reference subsequences from alignment data has been cumbersome and resource-intensive. This often involved intricate processes requiring the cross-referencing of multiple elements like query sequences, CIGAR strings, and MD tags. However, in recent years, a new tag called the CS tag has been developed [@Heng2018], which encapsulates the information contained in both CIGAR strings and MD tags. The CS tags enable researchers to represent mutations within alignments in a far more intuitive and streamlined manner.
+With the widespread commoditization of DNA sequencers, the scientific community is generating an unprecedented volume of sequence alignments on a daily basis. Conventionally, the task of identifying specific mutations or reconstructing reference subsequences from alignment data has been cumbersome and resource-intensive. This often involved intricate processes requiring the cross-referencing of multiple elements like query sequences, CIGAR strings, and MD tags. However, in recent years, a new tag called the cs tag has been developed [@Heng2018], which encapsulates the information contained in both CIGAR strings and MD tags. The cs tags enable researchers to represent mutations within alignments in a far more intuitive and streamlined manner.
 
-This paper presents two tools optimized for the use of CS tags: `cstag`, a Python library designed for effective manipulation and visualization of CS tags, and `cstag-cli`, a command-line utility for seamlessly appending CS tags to SAM/BAM files. By leveraging the representational strengths of CS tags, these tools simplify the extraction of critical mutation data from sequence alignments.
+This paper presents two tools optimized for the use of cs tags: `cstag`, a Python library designed for effective manipulation and visualization of cs tags, and `cstag-cli`, a command-line utility for seamlessly appending cs tags to SAM/BAM files. By leveraging the representational strengths of cs tags, these tools simplify the extraction of critical mutation data from sequence alignments.
 
 # Statement of Need
 
-The CS tag introduces a relatively new format for encoding information related to sequence alignments [@Heng2018]. While traditional formats, like CIGAR and MD tags, play crucial roles, the CS tag stands out by encoding sequence variants such as mismatches, insertions, and deletions all within one unified tag. To extract comprehensive variant data from CIGAR and MD tags, users often face the complexity of cross-referencing and interpreting between these tags, a task that can be intensive. On the other hand, the long form of the CS tag clearly indicates which bases from the query sequence align to specific bases in the reference sequence, improving its interpretability. As a result, the CS tag has simplified tasks that previously relied on mutual referencing between CIGAR, MD tags, and query sequences. This efficiency has led to the adoption of the CS tag in various bioinformatics tools, enhancing the process of variant identification and analysis [@Kuno2022; @Parker2021].
[…]
