+++ tests/test_to_vcf.py
@@ -79,7 +79,7 @@ def test_get_variant_annotations():
 
 
 ###########################################################
-# Format the CS tags
+# Format the cs tags
 ###########################################################
 
 
@@ -202,7 +202,7 @@ def test_add_vcf_fields():
 
 
 ###########################################################
-# process_cs_tag: Single CS tag
+# process_cs_tag: Single cs tag
 ###########################################################
 
 
@@ -227,7 +227,7 @@ chr1	5	.	C	CTT	.	.	."""
 
 
 ###########################################################
-# process_cs_tags: Multuple CS tags
+# process_cs_tags: Multuple cs tags
 ###########################################################
 
 
