### 2023-11-23 179e336
add `rm -rf /usr/local/bin/2to3` to avoid errors during `brew update` (https://github.com/actions/runner-images/issues/2322#issuecomment-749211076)
