+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "poetry.core.masonry.api"
 [tool.poetry]
 name = "cstag"
 version = "1.0.5"
-description = "Python module to manipulate the minimap2's CS tag"
+description = "Python library tailored for for manipulating and visualizing minimap2's cs tags"
 authors = ["Akihiro Kuno <akuno@md.tsukuba.ac.jp>"]
 homepage = "https://github.com/akikuno/cstag"
 repository = "https://github.com/akikuno/cstag"
