### 2023-11-23 f50fc70
add `pip install weasyprint` to avoid errors on `brew update`
