+++ .github/workflows/pytest.yml
@@ -23,15 +23,8 @@ jobs:
 
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip pytest
+          python -m pip install --upgrade pip pytest weasyprint
           python -m pip install -e .
-
-      - name: Install weasyprint on macos-latest
-        if: runner.os == 'macOS'
-        run: |
-          brew link --overwrite python@3.11
-          brew update
-          brew install weasyprint
 
       - name: Run pytest on Linux🐧, macOS🍎
         env:
