### 2023-11-23 1360a97
move top to install weasyprint to avoid conflicts of python env
