+++ .github/workflows/pytest.yml
@@ -3,7 +3,7 @@ name: Pytest
 on:
   push:
   schedule:
-    - cron: '0 2 * * 1' # The task runs at 2 a.m. every Monday.
+    - cron: "0 2 * * 1" # The task runs at 2 a.m. every Monday.
 
 jobs:
   build:
@@ -12,18 +12,24 @@ jobs:
       max-parallel: 4
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: [ '3.8', '3.9', '3.10', '3.11' ]
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
     name: Python ${{ matrix.python-version }} ${{ matrix.os }}
 
     steps:
       - uses: actions/checkout@v3
+
+      - name: Install weasyprint on macos-latest
+        if: runner.os == 'macOS'
+        run: |
+          brew install weasyprint
+
       - uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
 
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip pytest weasyprint
+          python -m pip install --upgrade pip pytest
           python -m pip install -e .
 
       - name: Run pytest on Linux🐧, macOS🍎
