### 2023-11-23 4979ecc
remove `brew update` to avoid errors on `brew update`
