+++ .github/workflows/pytest.yml
@@ -18,11 +18,6 @@ jobs:
     steps:
       - uses: actions/checkout@v3
 
-      - name: Install weasyprint on macos-latest
-        if: runner.os == 'macOS'
-        run: |
-          brew install weasyprint
-
       - uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
@@ -31,6 +26,11 @@ jobs:
         run: |
           python -m pip install --upgrade pip pytest
           python -m pip install -e .
+
+      - name: Install weasyprint on macos-latest
+        if: runner.os == 'macOS'
+        run: |
+          brew install weasyprint
 
       - name: Run pytest on Linux🐧, macOS🍎
         env:
