### 2023-11-23 akikuno/cstag#25 @akikuno
@betteridiot

Thank you very much for your invaluable review!

I have implemented the change to lowercase 'cs', as you pointed out, in both the paper and the GitHub documentation (README and Docstring).

JOSS Paper:

https://github.com/akikuno/cstag/commit/d7c6cfbad79b82ed1fa87d3a38e87e2c19768aaa

Document (README and Docstring):
https://github.com/akikuno/cstag/commit/20dec8d8d83d68f07fa13d1e89cfaf7cd5fe751d

I would appreciate your confirmation of these changes.
