### 2023-11-23 akikuno/cstag#25 @betteridiot
This satisfies my suggestion. Thank you.
