### 2023-11-24 2bfc6e9
add CI to Doctest
