+++ .github/workflows/doctest.yml
@@ -1,0 +1,39 @@
+name: Doctest
+
+on:
+  push:
+  schedule:
+    - cron: "0 2 * * 1" # The task runs at 2 a.m. every Monday.
+
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      max-parallel: 2
+      matrix:
+        os: [ubuntu-latest, macos-latest]
+        python-version: ["3.x"]
+    name: Python ${{ matrix.python-version }} ${{ matrix.os }}
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip pytest
+          python -m pip install -e .
+
+      - name: Install weasyprint on macos-latest
+        if: runner.os == 'macOS'
+        run: |
+          brew install weasyprint
+
+      - name: Run pytest on Linux🐧, macOS🍎
+        env:
+          PYTHONPATH: src
+        run: |
+          python -m pytest --doctest-modules
