### 2023-11-24 d37a56a
update docstrings to pass the doctest
