+++ src/cstag/shorten.py
@@ -14,7 +14,7 @@ def shorten(cs_tag: str, prefix: bool = False) -> str:
         >>> import cstag
         >>> cs = "=ACGT*ag=CGT"
         >>> cstag.shorten(cs, prefix=True)
-        cs:Z::4*ag:3
+        'cs:Z::4*ag:3'
     """
     cstags = re.split(r"([-+*~=])", cs_tag.replace("cs:Z:", ""))[1:]
     cstags = [i + j for i, j in zip(cstags[0::2], cstags[1::2])]
