### 2023-11-24 akikuno/cstag#26 @akikuno
@betteridiot

I sincerely thank you for your comments regarding Doctests.

I have made modifications to ensure that the Doctests pass.
https://github.com/akikuno/cstag/commit/d37a56adb69badbeb48993dd6b1d37484392738f

Additionally, I have added automatic testing for Doctests.
https://github.com/akikuno/cstag/actions/runs/6978602020/workflow#L39

I would be grateful if you could confirm them.
