### 2023-12-04 7545d62
fix the links at the attribute section of CODE_OF_CONDUCT.md
