### 2023-12-04 ed3f7c4
highlight `cstag-cli`
