### 2023-12-04 a4f516c
highlight `cstag-cli`
