### 2023-12-04 43cdbca
highlight `cstag-cli`
