### 2023-12-04 14f4b48
revise the Availability section to clearly state that cstag and cstag-cli are installed from a distinct repository
