### 2023-12-05 #6066 @akikuno
@jmschrei
Thank you very much for your careful coordination of the peer review process.

@betteridiot @jbloom
I deeply appreciate your excellent peer review comments.
Thanks to your support, the update of the source code is mostly complete (I will soon respond to @jbloom 's request at https://github.com/akikuno/cstag/issues/24), and I now plan to revise the JOSS documentation.

Let me comment here on the parts that have been revised.

@jbloom

## Could be helpful if it was a little bit more clearly indicated that cstag-cli is installed from a separate repo.

Thank you very much for your valuable comment.
In response, I have added to the *Availability* section of the main text that it is managed in a separate repository and registered separately in PyPI and BioConda.
https://github.com/akikuno/cstag/commit/14f4b4860e07f6c0acb0d09687313910b57fa2cf

Furthermore, in the README.md of cstag, I have incorporated a [NOTE] highlight as follows to indicate that cstag-cli is managed in a separate repository:

>[!NOTE]
> To add cs tags to SAM/BAM files, check out [`cstag-cli`](https://github.com/akikuno/cstag-cli).

https://github.com/akikuno/cstag#cstag

I would appreciate it if you could check this.
