### 2023-12-25 #6066 @akikuno
@betteridiot
Thank you very much for your careful confirmation.

> (optional) deconflict calcs package by describing the "rebranding" in either the cstag-cli or on calcs
> That is, let users know which package is the one that will continue to be maintained in case of dependencies.

As you pointed out, `calcs` is the predecessor of `cstag-cli` and has ceased development. While `calcs` differs from `cstag-cli` in that it can add cs tags to SAM files without MD tags, from a computational speed perspective, it is recommended to use `samtools calmd` to add MD tags before using `cstag-cli`.

Therefore, the README in the GitHub repository has been updated as follows:
- The deprecation of `calcs` has been clearly stated in its README.md.
  [calcs is deprecated](https://github.com/akikuno/calcs?tab=readme-ov-file#calcs-is-deprecated)

- Additionally, the README.md of `cstag-cli` has been updated to recommend using `samtools calmd` for adding MD tags.
  [cstag-cli README.md](https://github.com/akikuno/cstag-cli/blob/main/README.md?plain=1#L53)

Happy holidays!
