### 2023-12-27 78ce817
wip: add the description of `alignparse` and functionality of `2passtools`
