+++ paper/paper.md
@@ -27,7 +27,7 @@ This paper presents two tools optimized for the use of cs tags: `cstag`, a Pytho
 
 # Statement of Need
 
-The cs tag introduces a relatively new format for encoding information related to sequence alignments [@Heng2018]. While traditional formats, like CIGAR and MD tags, play crucial roles, the cs tag stands out by encoding sequence variants such as mismatches, insertions, and deletions all within one unified tag. To extract comprehensive variant data from CIGAR and MD tags, users often face the complexity of cross-referencing and interpreting between these tags, a task that can be intensive. On the other hand, the long form of the cs tag clearly indicates which bases from the query sequence align to specific bases in the reference sequence, improving its interpretability. As a result, the cs tag has simplified tasks that previously relied on mutual referencing between CIGAR, MD tags, and query sequences. This efficiency has led to the adoption of the cs tag in various bioinformatics tools, enhancing the process of variant identification and analysis [@Kuno2022; @Parker2021].
+The cs tag introduces a relatively new format for encoding information related to sequence alignments [@Heng2018]. While traditional formats, like CIGAR and MD tags, play crucial roles, the cs tag stands out by encoding sequence variants such as mismatches, insertions, and deletions all within one unified tag. To extract comprehensive variant data from CIGAR and MD tags, users often face the complexity of cross-referencing and interpreting between these tags, a task that can be intensive. On the other hand, the long form of the cs tag clearly indicates which bases from the query sequence align to specific bases in the reference sequence, improving its interpretability. As a result, the cs tag has simplified tasks that previously relied on mutual referencing between CIGAR, MD tags, and query sequences. This efficiency has led to the adoption of the cs tag in various bioinformatics tools, enhancing the process of variant identification and analysis such as a consensus variant calling[@Kuno2022],and extraction of splice junctions[@Parker2021].
 
 
 Here, the author has developed `cstag` and `cstag-cli` to further use the expression of the cs tag. `cstag` is a Python toolkit designed for a range of operations and visualizations related to cs tags. In addition, `cstag-cli` serves as a command-line utility focused on appending cs tags to pre-existing SAM/BAM files. `cstag` is designed to split and process sequence variant information, generate its reverse complement, and create consensus sequences. Furthermore, `cstag` can convert the variant information represented by the cs tag into Variant Call Format (VCF) files. In addition, it can produce outputs in HTML/PDF formats. The VCF conversion facilitates visualization in genome browsers, such as IGV [@Robinson2011], while the HTML/PDF formats cater to the generation of publication-ready figures (\autoref{fig:cstag-figure}). To illustrate its practical application, the authors have demonstrated how variant information in samples can be elucidated by deriving a consensus from multiple cs tags produced via Nanopore target sequencing, and subsequently visualizing the results [@Kuno2022].
@@ -37,6 +37,8 @@ Here, the author has developed `cstag` and `cstag-cli` to further use the expres
 
 Additionally, `cstag-cli` has been developed as a command-line tool specifically for appending cs tags to SAM/BAM files. While `paftools` [@Heng2018] can extract cs tags from these files, it converts them to the PAF format. Even though PAF is a lightweight and user-friendly format, the SAM/BAM format remains the predominant choice for many bioinformatics tools. Recognizing this preference, `cstag-cli` was designed to add cs tags directly to the original files without necessitating any format change. Accepting SAM/BAM files as input, `cstag-cli` outputs them with appended cs tags to the standard output. Given the tool's ability to directly process SAM/BAM files from standard input, it seamlessly integrates into existing scripts, enabling users to access cs tags without modifying their existing SAM/BAM file structures.
 
+
[…]
