### 2023-12-27 b844dda
Add description of "statement of need", comparison to `alignparse`, and acknowledge reviewer's comments.
