+++ paper/paper.md
@@ -27,26 +27,24 @@ This paper presents two tools optimized for the use of cs tags: `cstag`, a Pytho
 
 # Statement of Need
 
-The cs tag introduces a relatively new format for encoding information related to sequence alignments [@Heng2018]. While traditional formats, like CIGAR and MD tags, play crucial roles, the cs tag stands out by encoding sequence variants such as mismatches, insertions, and deletions all within one unified tag. To extract comprehensive variant data from CIGAR and MD tags, users often face the complexity of cross-referencing and interpreting between these tags, a task that can be intensive. On the other hand, the long form of the cs tag clearly indicates which bases from the query sequence align to specific bases in the reference sequence, improving its interpretability. As a result, the cs tag has simplified tasks that previously relied on mutual referencing between CIGAR, MD tags, and query sequences. This efficiency has led to the adoption of the cs tag in various bioinformatics tools, enhancing the process of variant identification and analysis such as a consensus variant calling[@Kuno2022],and extraction of splice junctions[@Parker2021].
+The cs tag introduces a relatively new format for encoding information related to sequence alignments [@Heng2018]. Unlike traditional formats, such as CIGAR and MD tags, the cs tag stands out by encoding sequence variants, including mismatches, insertions, and deletions, within a single, unified tag. Extracting comprehensive variant data from CIGAR and MD tags often requires users to navigate the complexity of cross-referencing and interpreting between these tags, a task that can be labor-intensive. In contrast, the long form of the cs tag clearly delineates how bases from the query sequence align with specific bases in the reference sequence, thus enhancing its interpretability. Consequently, the cs tag has simplified tasks that previously depended on mutual referencing between CIGAR, MD tags, and query sequences. This increased efficiency has led to the cs tag's adoption in various bioinformatics tools, facilitating processes like consensus variant calling [@Kuno2022] and filtering splice junctions [@Parker2021].
 
-
-Here, the author has developed `cstag` and `cstag-cli` to further use the expression of the cs tag. `cstag` is a Python toolkit designed for a range of operations and visualizations related to cs tags. In addition, `cstag-cli` serves as a command-line utility focused on appending cs tags to pre-existing SAM/BAM files. `cstag` is designed to split and process sequence variant information, generate its reverse complement, and create consensus sequences. Furthermore, `cstag` can convert the variant information represented by the cs tag into Variant Call Format (VCF) files. In addition, it can produce outputs in HTML/PDF formats. The VCF conversion facilitates visualization in genome browsers, such as IGV [@Robinson2011], while the HTML/PDF formats cater to the generation of publication-ready figures (\autoref{fig:cstag-figure}). To illustrate its practical application, the authors have demonstrated how variant information in samples can be elucidated by deriving a consensus from multiple cs tags produced via Nanopore target sequencing, and subsequently visualizing the results [@Kuno2022].
[…]
