### 2024-01-06 #6066 @akikuno
Happy New Year! May this year bring happiness and abundance.

@jbloom @betteridiot
In response to your comments, I've updated the paper with the following additions and modifications, and I hope these changes align with your expectations.

## [Added information and comparison about alignparse](https://github.com/akikuno/cstag/commit/b844ddaadf974bb2a70f85ea6a6603a563baebec#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424R36)

> Similar to `cstag`, the tool `alignparse` is also available [@Crawford2019]. In particular, both tools provide similar functions for formatting and converting cs tags, such as `cstag.split()` and `alignparse.cs_tag.split_cs()`. Comparing the two, `cstag` is relatively high-level, providing users with more direct functions tailored to specific needs. Examples of such functions include `cstag.to_vcf()` for VCF conversion and `cstag.to_pdf()` for visualization. In contrast, `alignparse` offers a broader range of functionalities beyond merely parsing cs tags. It is capable of tasks such as visualizing Genbank Flat files and aligning FASTQ files. Depending on their specific use case, users can choose between `cstag` and `alignparse`, or they might use them in conjunction for a more comprehensive analysis.

## [Added a sentence about the function of 2passtools (filtering splice junctions)](https://github.com/akikuno/cstag/commit/b844ddaadf974bb2a70f85ea6a6603a563baebec#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424R30)

>This increased efficiency has led to the cs tag's adoption in various bioinformatics tools, facilitating processes like consensus variant calling [@Kuno2022] and filtering splice junctions [@Parker2021].

## [Added a sentence to express statement of need at the State of Need section](https://github.com/akikuno/cstag/commit/b844ddaadf974bb2a70f85ea6a6603a563baebec#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424R32)
> Despite the growing popularity of the cs tag, there remains a lack of tools for effectively manipulating and visualizing it. As a result, users often have to create their own scripts for extracting and visualizing cs tags. In response, the author developed `cstag`. `cstag` is a Python toolkit for a range of operations and visualizations related to cs tags.

@jmschrei
I would like to update the version of cstag from v1.0.4 at the time of the JOSS submission to the current v1.0.5. How should I go about this?
