### 2024-01-15 #6066 @akikuno
I have updated the installation instructions for `cstag-cli` to accommodate Apple Silicon users by including the following note:

> [!NOTE]
> To Apple Silicon (ARM64) users:
> [Since the Bioconda channel does not yet support Apple Silicon](https://github.com/bioconda/bioconda-recipes/issues/37068#issuecomment-1257790919), please use the following command to install `cstag-cli` through Rosetta:
> ```bash
> CONDA_SUBDIR=osx-64 conda create -n env-cstag -c conda-forge -c bioconda python=3.10 cstag-cli -y
> conda activate env-cstag
> conda config --env --set subdir osx-64
> ```

https://github.com/akikuno/cstag-cli?tab=readme-ov-file#installation
