### 2024-01-15 #6066 @jmschrei
@jbloom @betteridiot would you mind commenting on the changes when you have a chance?

I'll change the version here. Just make sure that when you upload the code to a service that provides a DOI (like Zenodo) that it's the version you'd like distributed.
