### 2024-01-16 #6066 @akikuno
@jmschrei

Thank you so much!
I have completed the `Additional Author Tasks After Review is Complete`, including `the final version number for the software and Zenodo DOI containing all the relevant information (e.g., version, license, etc)`.
I hope it meets your expectations.

  - `cstag` version 1.0.5
  - `cstag-cli` version 1.0.0
  - `cstag` version 1.0.5 DOI: [10.5281/zenodo.8415528](https://doi.org/10.5281/zenodo.8415528)
  - `cstag-cli` version 1.0.0 DOI: [10.5281/zenodo.8358730](https://doi.org/10.5281/zenodo.8358730)
