### 2024-01-21 #6066 @Kevin-Mattheus-Moerman
@akikuno as the AEiC of this track I'll now help to process final steps. Since this review features two packages/repositories, and two version tags etc. Some changes are required:

- [ ] Please can you create a single ZENODO archive for the two software entities (matching the above tagged releases) and please report back here with that archive link.
- [ ] Ensure that all the metrics, like before, the title, the authors, the license matches. Next use `1.0.5, 1.0.0` as the version tag for the archive.
