### 2024-01-21 #6066 @Kevin-Mattheus-Moerman
@jmschrei note you did a good job here. Handling submissions featuring multiple repositories isn't streamlined yet I think, and we may optimise/formalise this better in the future. (I'll propose having comma-seperated repo/arhive/version tags etc for such papers, but this is not supported yet).
