### 2024-01-22 0dd5b5b
Add JOSS publication
